move core.model pytorch_model
